octeontx2-pf: Increase the size of dmacflt_max_flows
commit 23f1cf5c963dc7a1d18344e0453c27eeb2459f71 from
git@git.assembla.com:cavium/WindRiver.linux.git

 cn10kb supports large number of dmac filter flows to be
 inserted. Increase the field size to accommodate the same

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/98101
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_common.h
@@ -353,7 +353,7 @@ struct otx2_flow_config {
 #define OTX2_VF_VLAN_RX_INDEX	0
 #define OTX2_VF_VLAN_TX_INDEX	1
 	u16			max_flows;
-	u8			dmacflt_max_flows;
+	u32			dmacflt_max_flows;
 	u32			*bmap_to_dmacindex;
 	unsigned long		*dmacflt_bmap;
 	struct list_head	flow_list;
